+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Features
+- **responder:** http.NoBody now no longer set default status code to http.StatusNoContent
 - **responder:** added special handling for nil and http.NoBody
 
 ### Miscellaneous